Factorise out Contacts module.
