+++ src/ios/Diagnostic.h
@@ -13,7 +13,6 @@
 #import <CoreMotion/CoreMotion.h>
 #import <EventKit/EventKit.h>
 #import <AddressBook/AddressBook.h>
-#import <Contacts/Contacts.h>
 
 #import <mach/machine.h>
 #import <sys/types.h>
@@ -54,10 +53,6 @@ extern NSString*const AUTHORIZATION_GRANTED;
 - (void) enableDebug: (CDVInvokedUrlCommand*)command;
 
 - (void) switchToSettings: (CDVInvokedUrlCommand*)command;
-
-- (void) getAddressBookAuthorizationStatus: (CDVInvokedUrlCommand*)command;
-- (void) isAddressBookAuthorized: (CDVInvokedUrlCommand*)command;
-- (void) requestAddressBookAuthorization: (CDVInvokedUrlCommand*)command;
 
 - (void) getCalendarAuthorizationStatus: (CDVInvokedUrlCommand*)command;
 - (void) isCalendarAuthorized: (CDVInvokedUrlCommand*)command;
