+++ src/ios/Diagnostic_Contacts.h
@@ -1,0 +1,21 @@
+/*
+ *  Diagnostic_Contacts.h
+ *  Diagnostic Plugin - Contacts Module
+ *
+ *  Copyright (c) 2018 Working Edge Ltd.
+ *  Copyright (c) 2012 AVANTIC ESTUDIO DE INGENIEROS
+ */
+
+#import <Cordova/CDV.h>
+#import <Cordova/CDVPlugin.h>
+#import "Diagnostic.h"
+
+#import <Contacts/Contacts.h>
+
+@interface Diagnostic_Contacts : CDVPlugin
+
+- (void) getAddressBookAuthorizationStatus: (CDVInvokedUrlCommand*)command;
+- (void) isAddressBookAuthorized: (CDVInvokedUrlCommand*)command;
+- (void) requestAddressBookAuthorization: (CDVInvokedUrlCommand*)command;
+
+@end
