+++ www/ios/diagnostic.contacts.js
@@ -1,0 +1,98 @@
+/* globals cordova, require, exports, module */
+
+/**
+ *  Diagnostic Contacts plugin for iOS
+ *
+ *  Copyright (c) 2015 Working Edge Ltd.
+ *  Copyright (c) 2012 AVANTIC ESTUDIO DE INGENIEROS
+ **/
+var Diagnostic_Contacts = (function(){
+    /***********************
+     *
+     * Internal properties
+     *
+     *********************/
+    var Diagnostic_Contacts = {};
+
+    var Diagnostic = require("cordova.plugins.diagnostic.Diagnostic");
+
+    /********************
+     *
+     * Public properties
+     *
+     ********************/
+    
+    /********************
+     *
+     * Internal functions
+     *
+     ********************/
+
+    /*****************************
+     *
+     * Protected member functions
+     *
+     ****************************/
+
+
+    /**********************
+     *
+     * Public API functions
+     *
+     **********************/
+
+    /**
+     * Checks if the application is authorized to use contacts (address book).
+     *
+     * @param {Function} successCallback - The callback which will be called when operation is successful.
+     * This callback function is passed a single boolean parameter which is TRUE if contacts is authorized for use.
+     * @param {Function} errorCallback -  The callback which will be called when operation encounters an error.
+     * This callback function is passed a single string parameter containing the error message.
+     */
+    Diagnostic_Contacts.isContactsAuthorized = function(successCallback, errorCallback) {
+        return cordova.exec(Diagnostic._ensureBoolean(successCallback),
+            errorCallback,
+            'Diagnostic_Contacts',
+            'isAddressBookAuthorized',
+            []);
+    };
+
+    /**
+     * Returns the contacts (address book) authorization status for the application.
+     *
+     * @param {Function} successCallback - The callback which will be called when operation is successful.
+     * This callback function is passed a single string parameter which indicates the authorization status as a constant in `cordova.plugins.diagnostic.permissionStatus`.
+     * @param {Function} errorCallback -  The callback which will be called when operation encounters an error.
+     * This callback function is passed a single string parameter containing the error message.
+     */
+    Diagnostic_Contacts.getContactsAuthorizationStatus = function(successCallback, errorCallback) {
+        return cordova.exec(successCallback,
+            errorCallback,
+            'Diagnostic_Contacts',
+            'getAddressBookAuthorizationStatus',
+            []);
+    };
+
+    /**
+     * Requests contacts (address book) authorization for the application.
+     * Should only be called if authorization status is NOT_REQUESTED. Calling it when in any other state will have no effect.
+     *
+     * @param {Function} successCallback - The callback which will be called when operation is successful.
+     * This callback function is passed a single string parameter indicating whether access to contacts was granted or denied:
+     * `cordova.plugins.diagnostic.permissionStatus.GRANTED` or `cordova.plugins.diagnostic.permissionStatus.DENIED`
+     * @param {Function} errorCallback -  The callback which will be called when operation encounters an error.
+     * This callback function is passed a single string parameter containing the error message.
+     */
+    Diagnostic_Contacts.requestContactsAuthorization = function(successCallback, errorCallback) {
+        return cordova.exec(function(isGranted){
+                successCallback(isGranted ? Diagnostic.permissionStatus.GRANTED : Diagnostic.permissionStatus.DENIED);
+            },
+            errorCallback,
+            'Diagnostic_Contacts',
+            'requestAddressBookAuthorization',
+            []);
+    };
+
+    return Diagnostic_Contacts;
+});
+module.exports = new Diagnostic_Contacts();
